Flo's Flowers
I created this website during college as a final project. This was my first website to create with HTML and CSS without any additional help. It is still one of my favorite websites I have built to this day. One of my goals is to make it more professional with the store, adding to cart, and checking out page.